requirements.txt
APScheduler==3.9.1
asgiref==3.4.1
backports.zoneinfo==0.2.1
CacheControl==0.12.11
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.0.12
Django==3.2.2
django-cors-headers==3.10.1
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
httplib2==0.20.4
idna==3.3
importlib-resources==5.4.0
msgpack==1.0.4
mysqlclient==2.0.1
Pillow==8.4.0
proto-plus==1.20.6
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyotp==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
